Version 0.2.0.
Realised that GraphMap was more suited to my intended functionality so converted the Graph usage to GraphMap. Also added a trait to allow you to search for a key in a graph when you only know one character of it. This allows for easy searching of characters with shifting that differs across locales