Added tests, documentation and minor tweaks.
Minor tweak is that find key now no longer finds null characters as these should never occur on a keyboard currently in production and is used to show an unused state of a key.

All functions, traits, structs and enums now have doc comments.

Tests for adding alphabet and numeric characters to the graph as well as some testing on find_key with the resultant graphs.