Updated version number. Added example. Fixed bug.
When I did cargo fmt and the line breaks in the keyboard strings were removed when I fixed it the trailing '\' wasn't present resulting in an illformed input to the connect node function. This has been patched and a proper set of tests will be coming to ensure that this doesn't happen again.