Yet another coveralls/kcov fix..
Found one source that says you need a loop if there are multiple <PROJECT_NAME>-* hashes, otherwise not all your tests will be ran. Not sure if this will merge the single commit to a single statistic. If not I'll need to find a command to merge all the files and upload at once.